move the UDAnswerController to webcommon and have dataset summaries table in genomics sorting by default on create like dexp
