+++ packages/sites/genomics-site/webapp/wdkCustomization/js/client/component-wrappers/AnswerController.tsx
@@ -1,5 +1,7 @@
 import React, { ComponentType, Suspense, useMemo } from 'react';
 
+import UDAnswerController from '@veupathdb/web-common/lib/component-wrappers/UDAnswerController';
+
 import { Loading } from '@veupathdb/wdk-client/lib/Components';
 import { Props } from '@veupathdb/wdk-client/lib/Controllers/AnswerController';
 import { RecordInstance } from '@veupathdb/wdk-client/lib/Utils/WdkModel';
@@ -32,6 +34,11 @@ export function AnswerController(
           />
         ) : props.ownProps.recordClass === 'dataset' ? (
           <DatasetAnswerController
+            {...props}
+            DefaultComponent={DefaultComponent}
+          />
+        ) : props.ownProps.recordClass === 'userdataset' ? (
+          <UDAnswerController
             {...props}
             DefaultComponent={DefaultComponent}
           />
